requirements.txt
# FastAPI + ASGI
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Pydantic
pydantic==2.10.4
pydantic-settings==2.7.1

# LangGraph
langgraph==0.2.60
langchain-core==0.3.28
langchain-community==0.3.14

# LLM Providers
groq==0.13.1
google-generativeai==0.8.4
ollama==0.4.4

# Database — Supabase
supabase==2.13.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Auth — OAuth2
authlib==1.4.1
itsdangerous==2.2.0

# Task Queue — Celery + Redis
celery[redis]==5.4.0
redis==5.2.1

# HTTP Client
httpx==0.28.1

# Email Providers
google-api-python-client==2.156.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
msgraph-core==1.0.2

# Scraping + RSS
beautifulsoup4==4.12.3
feedparser==6.0.11
lxml==5.3.0

# Logging
structlog==24.4.0

# Embeddings
sentence-transformers==3.3.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Utilities
python-dateutil==2.9.0
python-dotenv==1.0.1
tenacity==9.0.0